Extract output container renderer
