+++ src/vs/workbench/contrib/notebook/browser/view/renderers/codeCell.ts
@@ -4,37 +4,31 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as DOM from 'vs/base/browser/dom';
-import { renderMarkdown } from 'vs/base/browser/markdownRenderer';
 import { raceCancellation } from 'vs/base/common/async';
 import { CancellationTokenSource } from 'vs/base/common/cancellation';
-import { IMarkdownString } from 'vs/base/common/htmlContent';
-import { Disposable, DisposableStore } from 'vs/base/common/lifecycle';
+import { Disposable } from 'vs/base/common/lifecycle';
 import { IDimension } from 'vs/editor/common/editorCommon';
-import { format } from 'vs/base/common/jsonFormatter';
-import { applyEdits } from 'vs/base/common/jsonEdit';
 import { IModeService } from 'vs/editor/common/services/modeService';
 import { IOpenerService } from 'vs/platform/opener/common/opener';
 import { IQuickInputService } from 'vs/platform/quickinput/common/quickInput';
 import { EDITOR_BOTTOM_PADDING, EDITOR_TOP_PADDING } from 'vs/workbench/contrib/notebook/browser/constants';
 import { CellFocusMode, CodeCellRenderTemplate, INotebookEditor } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
 import { CodeCellViewModel } from 'vs/workbench/contrib/notebook/browser/viewModel/codeCellViewModel';
-import { CellOutputKind, IInsetRenderOutput, IProcessedOutput, NotebookCellOutputsSplice, RenderOutputType } from 'vs/workbench/contrib/notebook/common/notebookCommon';
 import { INotebookService } from 'vs/workbench/contrib/notebook/common/notebookService';
 import { ITextFileService } from 'vs/workbench/services/textfile/common/textfiles';
 import { ClickTargetType } from 'vs/workbench/contrib/notebook/browser/view/renderers/cellWidgets';
-import { OutputElement } from 'vs/workbench/contrib/notebook/browser/view/renderers/cellOutput';
+import { OutputContainer } from 'vs/workbench/contrib/notebook/browser/view/renderers/cellOutput';
 
-const OUTPUT_COUNT_LIMIT = 500;
 
 export class CodeCell extends Disposable {
-	private outputEntries = new Map<IProcessedOutput, OutputElement>();
+	private _outputContainerRenderer: OutputContainer;
 
 	constructor(
 		private notebookEditor: INotebookEditor,
 		private viewCell: CodeCellViewModel,
 		private templateData: CodeCellRenderTemplate,
-		@INotebookService private notebookService: INotebookService,
-		@IQuickInputService private readonly quickInputService: IQuickInputService,
+		@INotebookService notebookService: INotebookService,
+		@IQuickInputService quickInputService: IQuickInputService,
 		@IOpenerService readonly openerService: IOpenerService,
 		@ITextFileService readonly textFileService: ITextFileService,
 		@IModeService private readonly _modeService: IModeService
@@ -142,17 +136,6 @@ export class CodeCell extends Disposable {
 			}
 		}));
 
-		this._register(viewCell.onDidChangeLayout(() => {
-			this.outputEntries.forEach((value, key) => {
-				const index = viewCell.outputs.indexOf(key);
-				if (index >= 0) {
-					const top = this.viewCell.getOutputOffsetInContainer(index);
-					value.domNode.style.top = `${top}px`;
-				}
-			});
-
-		}));
-
 		// Apply decorations
 		this._register(viewCell.onCellDecorationsChanged((e) => {
 			e.added.forEach(options => {
@@ -225,184 +208,10 @@ export class CodeCell extends Disposable {
 		updateFocusMode();
 
 		// Render Outputs
-		this.renderOutputs(editorHeight);
-		this._register(viewCell.onDidChangeOutputs((splices) => {
-			this.updateOutputs(splices);
-		}));
+		this._outputContainerRenderer = new OutputContainer(notebookEditor, viewCell, templateData, notebookService, quickInputService, openerService, textFileService);
+		this._outputContainerRenderer.render(editorHeight);
 		// Need to do this after the intial renderOutput
 		updateForCollapseState();
-	}
-
-	renderOutputs(editorHeight: number) {
-		if (this.viewCell.outputs.length > 0) {
-			let layoutCache = false;
-			if (this.viewCell.layoutInfo.totalHeight !== 0 && this.viewCell.layoutInfo.editorHeight > editorHeight) {
-				layoutCache = true;
-				this.relayoutCell();
-			}
-
-			this.templateData.outputContainer!.style.display = 'block';
-			// there are outputs, we need to calcualte their sizes and trigger relayout
-			// @TODO@rebornix, if there is no resizable output, we should not check their height individually, which hurts the performance
-			const outputsToRender = this.viewCell.outputs.slice(0, Math.min(OUTPUT_COUNT_LIMIT, this.viewCell.outputs.length));
-			for (let index = 0; index < outputsToRender.length; index++) {
-				const currOutput = this.viewCell.outputs[index];
-
-				// always add to the end
-				this.renderOutput(currOutput, index, undefined);
-			}
-
-			this.viewCell.editorHeight = editorHeight;
-			if (this.viewCell.outputs.length > OUTPUT_COUNT_LIMIT) {
-				this.templateData.outputShowMoreContainer.style.display = 'block';
-				this.viewCell.updateOutputShowMoreContainerHeight(46);
-			}
-
-			if (layoutCache) {
-				this.relayoutCellDebounced();
-			} else {
-				this.relayoutCell();
-			}
-		} else {
-			// noop
-			this.viewCell.editorHeight = editorHeight;
-			this.relayoutCell();
-			this.templateData.outputContainer!.style.display = 'none';
-		}
-
-		this.templateData.outputShowMoreContainer.innerText = '';
-		this.templateData.outputShowMoreContainer.appendChild(this.generateShowMoreElement());
-		// this.templateData.outputShowMoreContainer.style.top = `${this.viewCell.layoutInfo.outputShowMoreContainerOffset}px`;
-
-		if (this.viewCell.outputs.length < OUTPUT_COUNT_LIMIT) {
-			this.templateData.outputShowMoreContainer.style.display = 'none';
-			this.viewCell.updateOutputShowMoreContainerHeight(0);
-		}
-	}
-
-	updateOutputs(splices: NotebookCellOutputsSplice[]) {
-		if (!splices.length) {
-			return;
-		}
-
-		const previousOutputHeight = this.viewCell.layoutInfo.outputTotalHeight;
-
-		if (this.viewCell.outputs.length) {
-			this.templateData.outputContainer!.style.display = 'block';
-		} else {
-			this.templateData.outputContainer!.style.display = 'none';
-		}
-
-		const reversedSplices = splices.reverse();
-
-		reversedSplices.forEach(splice => {
-			this.viewCell.spliceOutputHeights(splice[0], splice[1], splice[2].map(_ => 0));
-		});
-
-		const removedKeys: IProcessedOutput[] = [];
-
-		this.outputEntries.forEach((value, key) => {
-			if (this.viewCell.outputs.indexOf(key) < 0) {
-				// already removed
-				removedKeys.push(key);
-				// remove element from DOM
-				this.templateData?.outputContainer?.removeChild(value.domNode);
-				this.notebookEditor.removeInset(key);
-			}
-		});
-
-		removedKeys.forEach(key => {
-			this.outputEntries.get(key)?.dispose();
-			this.outputEntries.delete(key);
-		});
-
-		let prevElement: HTMLElement | undefined = undefined;
-		const outputsToRender = this.viewCell.outputs.slice(0, Math.min(OUTPUT_COUNT_LIMIT, this.viewCell.outputs.length));
-
-		outputsToRender.reverse().forEach(output => {
-			if (this.outputEntries.has(output)) {
-				// already exist
-				prevElement = this.outputEntries.get(output)!.domNode;
-				return;
-			}
-
-			// newly added element
-			const currIndex = this.viewCell.outputs.indexOf(output);
-			this.renderOutput(output, currIndex, prevElement);
-			prevElement = this.outputEntries.get(output)?.domNode;
-		});
-
-		if (this.viewCell.outputs.length > OUTPUT_COUNT_LIMIT) {
-			this.templateData.outputShowMoreContainer.style.display = 'block';
-			this.viewCell.updateOutputShowMoreContainerHeight(46);
-		} else {
-			this.templateData.outputShowMoreContainer.style.display = 'none';
-		}
-
-		const editorHeight = this.templateData.editor!.getContentHeight();
-		this.viewCell.editorHeight = editorHeight;
-
-		if (previousOutputHeight === 0 || this.viewCell.outputs.length === 0) {
-			// first execution or removing all outputs
-			this.relayoutCell();
-		} else {
-			this.relayoutCellDebounced();
-		}
-	}
-
-	generateShowMoreElement(): any {
-		const md: IMarkdownString = {
-			value: `There are more than ${OUTPUT_COUNT_LIMIT} outputs, [show more ...](command:workbench.action.openLargeOutput)`,
-			isTrusted: true,
-			supportThemeIcons: true
-		};
-
-		const element = renderMarkdown(md, {
-			actionHandler: {
-				callback: (content) => {
-					if (content === 'command:workbench.action.openLargeOutput') {
-						const content = JSON.stringify(this.viewCell.outputs.map(output => {
-							switch (output.outputKind) {
-								case CellOutputKind.Text:
-									return {
-										outputKind: 'text',
-										text: output.text
-									};
-								case CellOutputKind.Error:
-									return {
-										outputKind: 'error',
-										ename: output.ename,
-										evalue: output.evalue,
-										traceback: output.traceback
-									};
-								case CellOutputKind.Rich:
-									return {
-										data: output.data,
-										metadata: output.metadata
-									};
-							}
-						}));
-						const edits = format(content, undefined, {});
-						const metadataSource = applyEdits(content, edits);
-
-						return this.textFileService.untitled.resolve({
-							associatedResource: undefined,
-							mode: 'json',
-							initialValue: metadataSource
-						}).then(model => {
-							const resource = model.resource;
-							this.openerService.open(resource);
-						});
-					}
-
-					return;
-				},
-				disposeables: new DisposableStore()
-			}
-		});
-
-		element.classList.add('output-show-more');
-		return element;
 	}
 
 	private viewUpdate(): void {
@@ -417,43 +226,15 @@ export class CodeCell extends Disposable {
 		}
 	}
 
-	private viewUpdateShowOutputs(): void {
-		for (let index = 0; index < this.viewCell.outputs.length; index++) {
-			const currOutput = this.viewCell.outputs[index];
-
-			const renderedOutput = this.outputEntries.get(currOutput);
-			if (renderedOutput && renderedOutput.renderResult) {
-				if (renderedOutput.renderResult.type !== RenderOutputType.None) {
-					this.notebookEditor.createInset(this.viewCell, renderedOutput.renderResult as IInsetRenderOutput, this.viewCell.getOutputOffset(index));
-				} else {
-					// Anything else, just update the height
-					this.viewCell.updateOutputHeight(index, renderedOutput.domNode.clientHeight);
-				}
-			} else {
-				// Wasn't previously rendered, render it now
-				this.renderOutput(currOutput, index);
-			}
-		}
-
-		this.relayoutCell();
-	}
-
 	private viewUpdateInputCollapsed(): void {
 		DOM.hide(this.templateData.cellContainer);
 		DOM.hide(this.templateData.runButtonContainer);
 		DOM.show(this.templateData.collapsedPart);
 		DOM.show(this.templateData.outputContainer);
 		this.templateData.container.classList.toggle('collapsed', true);
-
-		this.viewUpdateShowOutputs();
+		this._outputContainerRenderer.viewUpdateShowOutputs();
 
 		this.relayoutCell();
-	}
-
-	private viewUpdateHideOuputs(): void {
-		for (const e of this.outputEntries.keys()) {
-			this.notebookEditor.hideInset(e);
-		}
 	}
 
 	private viewUpdateOutputCollapsed(): void {
@@ -462,7 +243,7 @@ export class CodeCell extends Disposable {
 		DOM.show(this.templateData.collapsedPart);
 		DOM.hide(this.templateData.outputContainer);
 
-		this.viewUpdateHideOuputs();
+		this._outputContainerRenderer.viewUpdateHideOuputs();
 
 		this.templateData.container.classList.toggle('collapsed', false);
 		this.templateData.container.classList.toggle('output-collapsed', true);
@@ -477,11 +258,7 @@ export class CodeCell extends Disposable {
 		DOM.hide(this.templateData.outputContainer);
 		this.templateData.container.classList.toggle('collapsed', true);
 		this.templateData.container.classList.toggle('output-collapsed', true);
-
-		for (const e of this.outputEntries.keys()) {
-			this.notebookEditor.hideInset(e);
-		}
-
+		this._outputContainerRenderer.viewUpdateHideOuputs();
 		this.relayoutCell();
 	}
 
@@ -492,9 +269,7 @@ export class CodeCell extends Disposable {
 		DOM.show(this.templateData.outputContainer);
 		this.templateData.container.classList.toggle('collapsed', false);
 		this.templateData.container.classList.toggle('output-collapsed', false);
-
-		this.viewUpdateShowOutputs();
-
+		this._outputContainerRenderer.viewUpdateShowOutputs();
 		this.relayoutCell();
 	}
 
@@ -516,12 +291,7 @@ export class CodeCell extends Disposable {
 		);
 
 		// for contents for which we don't observe for dynamic height, update them manually
-		this.viewCell.outputs.forEach((o, i) => {
-			const renderedOutput = this.outputEntries.get(o);
-			if (renderedOutput && renderedOutput.renderResult && renderedOutput.renderResult.type === RenderOutputType.None && !renderedOutput.renderResult.hasDynamicHeight) {
-				this.viewCell.updateOutputHeight(i, renderedOutput.domNode.clientHeight);
-			}
-		});
+		this._outputContainerRenderer.onCellWidthChange();
 	}
 
 	private onCellHeightChange(newHeight: number): void {
@@ -534,14 +304,6 @@ export class CodeCell extends Disposable {
 				height: newHeight
 			}
 		);
-	}
-
-	private renderOutput(currOutput: IProcessedOutput, index: number, beforeElement?: HTMLElement) {
-		if (!this.outputEntries.has(currOutput)) {
-			this.outputEntries.set(currOutput, new OutputElement(this.notebookEditor, this.notebookService, this.quickInputService, this.viewCell, this.templateData.outputContainer, currOutput));
-		}
-
-		this.outputEntries.get(currOutput)!.render(index, beforeElement);
 	}
 
 	relayoutCell() {
@@ -567,10 +329,7 @@ export class CodeCell extends Disposable {
 
 	dispose() {
 		this.viewCell.detachTextEditor();
-		this.outputEntries.forEach((value) => {
-			value.dispose();
-		});
-
+		this._outputContainerRenderer.dispose();
 		this.templateData.focusIndicatorLeft!.style.height = 'initial';
 
 		super.dispose();
